# Doubled the speed of the ebrisk calculator