--- a/openquake/calculators/ebrisk.py
+++ b/openquake/calculators/ebrisk.py
@@ -89,7 +89,6 @@ def ebrisk(rupgetter, srcfilter, param, monitor):
         imts = getter.imts
         eids = rupgetter.get_eid_rlz()['eid']
         eid2idx = {eid: idx for idx, eid in enumerate(eids)}
-        rlz2idx = rupgetter.rlz2idx
         tagnames = param['aggregate_by']
         shape = assgetter.tagcol.agg_shape((len(eids), L), tagnames)
         acc = numpy.zeros(shape, F32)  # shape (E, L, T, ...)
@@ -109,8 +108,8 @@ def ebrisk(rupgetter, srcfilter, param, monitor):
                 tidx = tagidxs[asset.ordinal]
                 acc[(eidx, lti) + tidx] += losses
                 if param['avg_losses']:
-                    for eid, rlz, loss in zip(haz['eid'], haz['rlzi'], losses):
-                        losses_by_RN[rlz2idx[rlz], sid, lti] += loss
+                    losses_by_RN[:, sid, lti] += rupgetter.E2R(
+                        losses, haz['rlzi'])
             times[sid] = time.time() - t0
     return {'losses': acc, 'eids': eids, 'losses_by_RN':
             (rupgetter.rlzs, losses_by_RN), 'times': times}