Another step towards https://github.com/gem/oq-engine/issues/4390.
Before:
```
$ oq2 show performance 22315
============================== ======== ========= ======
operation                      time_sec memory_mb counts
============================== ======== ========= ======
total ebrisk                   180,235  310       208   
building risk                  170,169  36        208   
getting hazard                 10,032   307       208   
getting assets                 2,777    0.0       208   
saving losses_by_event         445      14        208   
```
After:
```
$ oq2 show performance 22316
============================== ======== ========= ======
operation                      time_sec memory_mb counts
============================== ======== ========= ======
total ebrisk                   97,050   389       208
building risk                  86,675   46        208
getting hazard                 10,341   373       208
getting assets                 2,794    0.0       208
saving losses_by_event         441      15        208
```

The speedup is in the calculation of the avg_losses.